+++ src/userio.cpp
@@ -13,11 +13,10 @@ void UserIO::setScale(sf::Shape &shape)
 {
     sf::Vector2u size{window_.getSize()};
     float xScaleFactor{size.x / 1000.f};
-    // float yScaleFactor{size.y / 1000.f};
     shape.scale(xScaleFactor, 1.0f);
 }
 
-void UserIO::drawOnScreen(Shapes::Type shapeIdx, float x, float y, float xScale, float yScale, float dim, int textureOffset)
+void UserIO::drawOnScreen(Shapes::Type shapeIdx, float x, float y, float xScale, float yScale, float dim, float textureOffset)
 {
     sf::Vector2u size{window_.getSize()};
     sf::RectangleShape shape{Shapes::getShape(shapeIdx, textureOffset)};
